Move VS2019/VS2022 shared code to SourceControlDLShared
