+++ SourceControlDLShared/Commands/DeepLinkFormatDynamicCommand.cs
