+++ SourceControlDLShared/Helpers/BitbucketProviderHelper.cs
