+++ SourceControlDLShared/Helpers/GitHelper.cs
