+++ SourceControlDLShared/Helpers/GithubProviderHelper.cs
