+++ SourceControlDLShared/Helpers/ProviderHelper.cs
