+++ SourceControlDLShared/Helpers/ProviderLinkInfo.cs
