+++ SourceControlDLShared/Options/ProviderInfo.cs
