+++ SourceControlDeepLinks/SourceControlDeepLinks.csproj
@@ -47,20 +47,7 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Commands\BitbucketDeepLinkCommand.cs" />
-    <Compile Include="Commands\BitbucketLinesDeepLinkCommand.cs" />
-    <Compile Include="Commands\DeepLinkFormatDynamicCommand.cs" />
-    <Compile Include="Helpers\BitbucketProviderHelper.cs" />
-    <Compile Include="Helpers\GitHelper.cs" />
-    <Compile Include="Helpers\GithubProviderHelper.cs" />
-    <Compile Include="Helpers\ProviderHelper.cs" />
-    <Compile Include="Helpers\ProviderLinkInfo.cs" />
-    <Compile Include="Options\ExtensionOptions.cs" />
-    <Compile Include="Options\IExtensionOptions.cs" />
-    <Compile Include="Options\ProviderInfo.cs" />
-    <Compile Include="Options\SourceProviderEnum.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Resources\Constants.cs" />
     <Compile Include="SourceControlDeepLinksPackage.cs" />
     <Compile Include="source.extension.cs">
       <AutoGen>True</AutoGen>
@@ -116,7 +103,13 @@
       <Version>4.5.4</Version>
     </PackageReference>
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Commands\" />
+    <Folder Include="Helpers\" />
+    <Folder Include="Options\" />
+  </ItemGroup>
   <Import Project="..\SharedSrc\SharedSrc.projitems" Label="Shared" />
+  <Import Project="..\SourceControlDLShared\SourceControlDLShared.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
