--- SourceControlDeepLinksVS2022/Options/SourceProviderEnum.cs
@@ -1,9 +1,0 @@
-﻿namespace SourceControlDeepLinks.Options
-{
-	public enum SourceProvider
-	{
-		// 0 reserved
-		BitbucketServer = 1,
-		GitHub = 2
-	}
-}
