--- SourceControlDeepLinksVS2022/Resources/Constants.cs
@@ -1,10 +1,0 @@
-﻿namespace SourceControlDeepLinks.Resources
-{
-	public static class Constants
-	{
-		public const string ExtensionOutputPane = "Source Control Deep Links";
-		public const string OptionsPageCategoryName = "Source Control Deep Links";
-		public const string OptionsPageName = "General";
-		public const string SourceLink = "Source Link";
-	}
-}
