+++ SourceControlDeepLinksVS2022/SourceControlDeepLinksVS2022.csproj
@@ -17,7 +17,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>SourceControlDeepLinks</RootNamespace>
-    <AssemblyName>SourceControlDeepLinks</AssemblyName>
+    <AssemblyName>SourceControlDeepLinks22</AssemblyName>
     <TargetFrameworkVersion>v4.8</TargetFrameworkVersion>
     <GeneratePkgDefFile>true</GeneratePkgDefFile>
     <UseCodebase>true</UseCodebase>
@@ -48,22 +48,8 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Commands\BitbucketDeepLinkCommand.cs" />
-    <Compile Include="Commands\BitbucketLinesDeepLinkCommand.cs" />
-    <Compile Include="Commands\DeepLinkFormatDynamicCommand.cs" />
-    <Compile Include="Helpers\BitbucketProviderHelper.cs" />
-    <Compile Include="Helpers\GitHelper.cs" />
-    <Compile Include="Helpers\GithubProviderHelper.cs" />
-    <Compile Include="Helpers\ProviderHelper.cs" />
-    <Compile Include="Helpers\ProviderLinkInfo.cs" />
-    <Compile Include="Helpers\SettingsHelper.cs" />
-    <Compile Include="Options\ExtensionOptions.cs" />
-    <Compile Include="Options\IExtensionOptions.cs" />
-    <Compile Include="Options\ProviderInfo.cs" />
-    <Compile Include="Options\SourceProviderEnum.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Commands\MyCommand.cs" />
-    <Compile Include="Resources\Constants.cs" />
     <Compile Include="SourceControlDeepLinksVS2022Package.cs" />
     <Compile Include="source.extension.cs">
       <AutoGen>True</AutoGen>
@@ -114,7 +100,12 @@
       <Version>6.0.0</Version>
     </PackageReference>
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Helpers\" />
+    <Folder Include="Options\" />
+  </ItemGroup>
   <Import Project="..\SharedSrc\SharedSrc.projitems" Label="Shared" />
+  <Import Project="..\SourceControlDLShared\SourceControlDLShared.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
